Module test_store::block_store

Re-exports

pub use graph_store_postgres::layout_for_tests::chain_support::SettableChainStore as _;

Structs

BLOCK_FIVE

BLOCK_FOUR

BLOCK_ONE

BLOCK_ONE_NO_PARENT

BLOCK_ONE_SIBLING

BLOCK_SIX_NO_PARENT

BLOCK_THREE

BLOCK_THREE_NO_PARENT

BLOCK_TWO

BLOCK_TWO_NO_PARENT

FakeBlock
The parts of an Ethereum block that are interesting for these tests: the block number, hash, and the hash of the parent block

GENESIS_BLOCK

Constants

NO_PARENT

Functions

set_chain
Store the given chain as the blocks for the network set the network’s genesis block to genesis_hash, and head block to null

Type Definitions

Chain